Make sure that a receiver is always a simple instance or klass or any
In short, `Array` is okay as a receiver, but `Array[Integer]` is not.
If we allow a type-applied class as a receiver, it very easily leads to
a state explosion because one method should be analyzed for
`Array[Integer]`, `Array[String]`, `Array[Array[Integer]]`, ...

Note, putself must not push `Array` directly to the stack.