Stop unnecessary env merging when a new callsite is found
Formerly, `add_callsite!` propagated a new return environments to all
callsites, but it was not necessary. `add_callsite!` only adds a new
callsite. It doesn't change return environments of other callsite than
the new one, doesn't change a return type. Thus, it is enough to
propagate a environment only to the new callsite.

This makes some kind of code much faster to analyze, such as
https://github.com/ruby/matrix/blob/c6043917b394056d121e97e9871d59ce48491f3e/lib/matrix.rb#L1312-L1324
See #14.