Show "Array[untyped]" when an empty array is guessed
Formerly, an empty array is shown as "Array[bot]".
It is representable as "[ ]" in RBS, but it is considered that an empty
array is a wrong guess due to the analysis limitation of TypeProf.
Typically, users often have to modify it to "Array[something]".
So, "Array[untyped]" is considered more useful than "[ ]".

Similarly, "Hash[bot, bot]" is changed to "Hash[untyped, untyped]".