Add $typeprof as always true for type annotation
This feature may very controversial, but it allows to force cast a type
as follows:

    x = 42
    x = "foo" if $typeprof
    p(x)  #=> String

Unfortunately, TypeProf often does wrong type inference, which is
unavoidable. However, there was no simple way to correct the inference.
To be honest I don't like this feature, but writing RBS may be too heavy
for correcting trivial misinferences.