Initial support lambda node
Prior to this change, typeprof would crash on encountering code like:

```rb
def x = -> { 1 }
```

This was because typeprof did not yet support lambda.
With this update, such lambda definitions are now recognized as `def x: -> Proc`.
While the ideal scenario would be to infer the lambda's type directly (e.g., `^() -> Integer` as in RBS),
this initial step addresses the crash issue and provides basic support.