Add settings for GitHub Codespaces
## Summary

GitHub Codespaces allows Ruby engineers to easily launch an environment to play with TypeProf.

This commit adds the files necessary to use GitHub Codespaces.

## Description

I created configuration files according to the documentation below.
refs: https://docs.github.com/en/codespaces/setting-up-your-project-for-codespaces/adding-a-dev-container-configuration/introduction-to-dev-containers#using-a-predefined-dev-container-configuration

I followed the steps in this doc and used Ruby from the list of predefined dev container configurations.
The changes made from the default settings are as follows.

- Add `rbs` because the `ruby:1-3.3-bookworm` image does not have rbs by default
- Add `mame.ruby-typeprof` extension and set `typeprof.server.path`
- Associate `typeprof.conf.json` with `jsonc`