Make constant resolution strict for Module#include
In Ruby, `include A; include B` tries to include `A::B` if any.
TypeProf tried to reproduce this behavior, but it turned out to be
difficult to keep it. This change makes constant resolution not respect
the inheritance for the arguments of Module#include.

TypeProf retried constant resolution whenever its constant scope is
changed, including the inheritance is changed. However, this retry
mechanism caused infinite loop when there is a constant like `A::A`.
`include A` tries to include the outer `A`, which changes the
inheritance, so the `include A` is reanalyzed, and then it tries to
include the inner `A::A`, which changes the inheritance again, and then
it tries to include the outer `A` again, which returns the first state.

By this change, `include A` no longer tries to include the inner `A::A`,
so it can avoid the infinite loop.

This fixes #285