Make `untyped` not match any other type
So far, we have assumed that `untyped` matches any type. This approach
had a problem that when an argument is untyped, it matches a method
signature, and when the argument is any concrete type, it does not match
the same method signature, and if the result of the match affects the
type of the argument itself, the analysis would oscillate and loop
endlessly.

This change makes `untyped` not match any other type. This will have a
significant difference from gradutal typing's `untyped` (or `any`).
Instead, we suppress diagnostics like `expected: C, actual: untyped`.
We expect that this will not make much difference on the user side.

By assuming that untyped does not match a concrete type, it is expected
that the scope of analysis will be narrower (not enough analysis can be
done after the method call that fails to resolve overload).
Therefore, we decided that if there is no overload in the method
signature, the method signature will always be chosen (errors will be
reported if there is a discrepancy in the type or number of arguments).
This is expected to allow many method calls to be parsed with concrete
types afterwards, even if there are discrepancies in arguments.