Fixed typeprof.conf.json could not be loaded on VSCode on Windows
The problem was that the `url` processed by `uri_to_path` was like `c%3A/Users/shunh/vscode-typeprof-test`, and typeprof.conf.json could not be loaded.

This meant that TypeProf could be launched as an LSP, but the results of type analysis were not displayed on VSCode.

So, I changed that if `%3A` is included in the url returned by `uri_to_path`, it is replaced with `:`.